Add runtime support for LIters
Summary:
Add support in the runtime for LIters. This includes the new bytecode
ops LIterInit, LIterInitK, LIterNext, LIterNextK, and LIterFree (LIter support
in IterBreak has already been added). These behave similarily to the normal
iterator ops, except they all take an associated local. Under normal conditions,
these iterators do not actually store their base. Instead they use the
associated local as the base. If the local contains an object, the iterator will
take ownership of the base. In that case the thing being iterated over is not
actually what's stored in the local (IE, an internal array from a collection),
so this optimization cannot apply.

The advantage of this is that we can avoid increasing the ref-count of the array
being iterated over, which means we can avoid COW in certain situations within
the loop. Its not safe to use LIters blindly. They're only safe to use under
certain conditions (the associated base local cannot change in a semantically
meaningful way while within the foreach loop). HHBBC is required to perform this
optimization, so there are no actual uses of these ops in this diff.

When emitting code, the JIT will use the known type of the associated local to
either lower the LIter* ops to their associated LIter HHIR instructions, or just
normal Iter operations. The criteria is as described above. If the local
contains an ArrLike, LIter HHIR instructions are used, normal HHIR instructions
for objects. This is safe because a pre-condition of these ops is that the
associated local can't change types while the iterator is alive.

Templatize some of the array-iterator runtime helpers to control whether they
take ownership (with associated ref-count manipulations) or not. Add support for
these ops in HHBBC, factoring out the iterator implementation into a helper
function.

Reviewed By: swtaarrs

Differential Revision: D7674985

fbshipit-source-id: b5d3c931f4667c5b394d7afe2c76d5960ec113d7