Add Let to AST
Summary:
Background: we want to add `let` statements to enable declaration & initialization of variables in the following ways:

```
let x = 3;
let x : int = 3;
```

This diff is the first step and adds a `Let` constructor to `stmt` in `ast.ml`. I've also updated any code that uses the AST.

For now, we're not doing any codegen or typechecking, but I've added TODOs linking to tasks.

Differential Revision: D7413933

fbshipit-source-id: 521937a58c5bd7d0565382346a69c8e24d419fd1