Make the environment provided by Tast_visitor abstract
Summary:
Currently, `Tast_visitor` provides a `Typing_env.env` to its visitor methods. This is a bit of a hazard to TAST consumers, since `Tast_visitor` is able to restore only a small subset of the typing environment from the TAST's `EnvAnnotation.t`. It is not obvious which functions that take a `Typing_env.env` are safe to call--many will return a subtly incorrect value.

This change adds a `Tast_env` module, which abstracts away the implementation of `Tast_visitor`'s environment type, and provides aliases and documentation for the functions that will return correct results on an environment restored from a TAST `EnvAnnotation.t`.

Reviewed By: pittsw

Differential Revision: D7710944

fbshipit-source-id: 8f5ec9187372be1f813cc29b947da2d8309834dc