Make Fallthrough behavior in FFP and legacy parser somewhat consistent
Summary:
Fallthroughs in the legacy parser could be put *anywhere* in the case block. In the FFP, this is much more correct, in that we only allow // FALLTHROUGH as the last statement in a case block.

Fallthroughs in legacy parser could have text after it, aka
// FALLTHROUGH something something
would be correct. This is good behavior, so I've made the FFP do that.
We need the two parsers to agree on what is valid fallthrough behavior here, in order to be push safe when releasing the FFP in www.

Reviewed By: kmeht, oulgen

Differential Revision: D7779773

fbshipit-source-id: e25c7800a2bd191d77fd8e504e70d14ab8ea57fc