Require a typehint when overriding typed members
Summary: We want to enforce property typehints in HHVM. In order to make this process safer, we should enforce that if a member is typed in a parent, it must also be typed in its children. This way, we can guarantee that its type remains the same through the class hierarchy.

Reviewed By: dlreeves

Differential Revision: D7780349

fbshipit-source-id: 42a859bbd8b368817e19f717cf45499c1db4aa18