Consolidated cut-over for parsing service and heap
Summary:
This is a copy of D7548168. It should be more correct now, since the diff below this uses HHVM compat mode for hh_server, meaning that only errors that would fatal at runtime would show as parse errors now.

Above this diff will be changes related to the *performance* of FFP in hh_server, but the *correctness* should be good at this point.

Differential Revision: D7767133

fbshipit-source-id: 821dfd1947d7c23e7605ec676b0d6cbb3bbe8cdd