Parsing and bytecode emission for __hhas_adata()
Summary:
**This diff introduces a special function, `__hhas_adata()`, which is a //temporary// workaround to allow the FFP to parse large arrays without running out of memory.**

autoload_map.php is a large `darray`:

```
$ cat /var/www/flib/autoload/autoload_map.php | \
  (head -n 15 && echo "..." && tail -n 7)
<?hh // decl
/**
 * autogenerated by Facts DB
 *
 * nogrep
 * build-artifact
 * do-not-extract
 */
function __flib_autoload_get_class_map() {
  return darray[
    '__configeratoruistore_filesaccessor' => 'flib/configerator/data/__tests__/ConfigeratorUIStoreTest.php',
    '__gkcheckdisplaytestvalidationdisplay' => 'flib/intern/gatekeeper/mutators/__tests__/GKPendingMutationsTestCase.php',
    '__gkchecktracewithcheckexceptionsclass' => 'flib/intern/gatekeeper/check_trace/__tests__/GKCheckTraceWithCheckExceptionsTest.php',
    '__gkchecktracewithcheckexceptionsclasstest' => 'flib/intern/gatekeeper/check_trace/__tests__/GKCheckTraceWithCheckExceptionsTest.php',
    '__gkchecktracewithconfigeratorstorageclass' => 'flib/intern/gatekeeper/check_trace/__tests__/GKCheckTraceWithConfigeratorStorageTest.php',
...
    'zplformirf' => 'flib/intern/site/x/zebra_printers/ZPLFormTypeEnum.php',
    'zplformlostphone' => 'flib/intern/site/x/zebra_printers/ZPLFormTypeEnum.php',
    'zplformtelcoreturn' => 'flib/intern/site/x/zebra_printers/ZPLFormTypeEnum.php',
    'zrtestipmetadata' => 'flib/mobile/zero/zrtest/ZRTestIPInfo.php',
    'zrtestipmetadatavector' => 'flib/mobile/zero/zrtest/ZRTestIPInfo.php',
  ];
}
```

The FFP takes too much memory parsing it. We want to reduce the memory that it takes to load FFP.

This problem exists because the FFP has bad memory performance when parsing lots of nested tokens, like in the massive `darray` seen above.

Since we are in control of the generation of the autoload map, we could generate it directly as `.adata`, which is what gets output in the hhas bytecode anyway. Imagine we had a magical function, let's call it `__hhas_adata()`, which allowed us to directly output `.adata` for HackC. With this, we could implement the above using:

```
function __flib_autoload_get_class_map() {
  // This is just an example.
  return __hhas_adata(
    'Y:4:{s:35:\"__configeratoruistore_filesaccessor\";s:60:\"flib/configerator/data/__tests__/ConfigeratorUIStoreTest.php\";s:37:\"__gkcheckdisplaytestvalidationdisplay\";s:72:\"flib/intern/gatekeeper/mutators/__tests__/GKPendingMutationsTestCase.php\";s:16:\"zrtestipmetadata\";s:40:\"flib/mobile/zero/zrtest/ZRTestIPInfo.php\";s:22:\"zrtestipmetadatavector\";s:40:\"flib/mobile/zero/zrtest/ZRTestIPInfo.php\";}'
  );
}
```

This diff implements that function. Obviously, this is a **short-term workaround** and will never see the light of day for general developers.

I introduced a new syntax for this special function, since it needs special handling throughout HackC. It is treated very similarly to `Array` and cousins. The `full_fidelity_ast` lowerer will reject any argument that is not a single- or double-quoted string -- this special function only works on string literal arguments.

For codegen, surprisingly little had to be done. The `Ast_constant_folder` already supports extracting `.adata` fields for arrays, and we just pretend this is an array for the most part. The one difference, of course, is that instead of iterating over the array's elements for `.adata` extraction, we simply dump the string literal as the `.adata` payload.

Reviewed By: paulbiss

Differential Revision: D7754665

fbshipit-source-id: e587052f8f9fb5a24c2be314a8f31be6c84388ff