Make thread counting more robust
Summary:
When running a dbg build under gdb, I get the

  HHVM is broken: %u threads running in hugifyText()

message about 1 out of every 3 runs. If I set a breakpoint on that
line, and then do `info thread` when it fires, gdb always tells me
there's only one thread; so there seems to be a race.

Try to fix it by sleeping and retrying.

Reviewed By: binliu19

Differential Revision: D7814401

fbshipit-source-id: 16f45e99394a748150d8bac2c9cecc15d2ceea29