Fix promotion of pending locals
Summary:
Consider the following example (minimized from the example in [[ https://fb.facebook.com/groups/hackforhiphop/permalink/1883671505014789/ | this ]] post):
```
<?hh // strict

function test1(): void {
  if (true) {
    if (false) {}
    // switch (false) {}
    // try {} catch (Exception $_) {}
    echo $x;
    $x = 1;
  } else {
    $x = 2;
  }
}
```
It type-checks even though clearly the variable `$x` in the then-branch is referenced before it is defined.  The same behavior is observed if you replace `if (false) {}` with any of the commented out statements.

The problem here is in how we deal with so called pending locals (e.g., two instances of `$x` defined in the two branches of a conditional).  Pending locals become defined only after the conditional, so they are handled as following :

diffusion/FBS/browse/master/fbcode/hphp/hack/src/naming/naming.ml;116f59942c31952c99dbfbe9c7aeaa68be76a9cb$1855-1869

Before processing the conditional, we extract all local variabless from it and introduce a pending local for each of them (so before processing the outer conditional in our example the map of pending locals contains a single entry keyed by `"$x"`).  We also stash away the locals and pending locals (that's what `Env.scope` does).  We then process the branches, restore the locals and pending locals stashed away previously (`Env.scope`), and promote each pending local to a proper local.

It's easy to see now what happens in the above example: the inner conditional is processed in an environment in which the pending locals contain `"$x"`, and after the branches of the inner conditional have been processed we restore the pending locals that we had upon entering it (which again contain `"$x"`), and then we promote //all// pending locals to regular locals.  I.e., we moved `"$x"` from pending to ordinary locals even though it hasn't been defined by the inner conditional.

We do this promotion of pending locals for switch and try statements, which suffer from a similar problem.  A fix I'm proposing is to promote only the pending locals that are contained in the if/switch/try statement we have just finished processing.

Reviewed By: fangyi-zhou

Differential Revision: D7815126

fbshipit-source-id: 9b002a39550e8589e886592fa49e42e04814bcba