Perform Iter to LIter optimization in HHBBC
Summary:
Add an optimization pass to HHBBC to turn Iter operations into LIter
operations. HHBBC will track an iterator's associated local (if any). If that
associated local is not changed on all paths from the initialization point to
all usages of that iterator, it is safe to turn all the ops for that iterator to
the LIter variants.

This current criteria is more strict than necessary, as certain kinds of writes
can be safely allowed. This will be loosened in a follow-on diff.

Reviewed By: markw65

Differential Revision: D7795598

fbshipit-source-id: 83c7b37a76dc626b5ca18b2e1a79a4074aaf3085