RFC: Sealed Classes
Summary:
This adds sealed classes to the typechecker, leaving interfaces for a followup diff. Here, we allow classes to specify a `__Sealed` attribute, which takes variadic classnames representing a whitelist of classes that can extend it. We also enforce that a class is not both sealed and final, since that's redundant.

With sealed classes, we can prevent builtin classes/interfaces like Traversable, Container, etc. from being extended/implemented in userland. This also sets a foundation for other features we may want to consider in the future, like ADTs and pattern matching.

Reviewed By: jamesjwu

Differential Revision: D7824439

fbshipit-source-id: bad4b60d045187e34dddb35f0aa7676d92575c92