Use switch table for mayRaiseError
Summary: A number of IR instructions set `Er` in ir.specification and then explicitly override it based on the values of the various hack array compat flags in `ir-instruction-inl.h` and `ir-opcodes-inl.h`. Instead lets use a switch table so that we can condition on runtime options. This cleanup should make some changing to error handling that I have planned easier.

Reviewed By: jano

Differential Revision: D7857355

fbshipit-source-id: 8ffa939acb9f27f717b366261654d14ec72c7101