Make type-checking of comparisons stricter
Summary:
We have known for a while that type-checking of comparisons is broken.  E.g., the following example type-checks:
```
<?hh // strict

function id<T>(T $x): T {
  return $x;
}

class A {}
class B {}

function test(): bool {
  $x = id(new A());
  $y = id(new B());
  return $x > $y;
}
```
That's because application of `id()` wraps `A` and `B` into unresolved, so that the types of `$x` and `$y` becomes `(A as A)` and `(B as B)` respectively, and they happily unify (we allow comparisons when the types of the left and right hand side unify).  Also, the presence of unresolved allows one to compare against nullable values:
```
<?hh // strict

function id<T>(T $x): T {
  return $x;
}

function test(int $x): bool {
  return id($x) < null;
}
```
(which likely doesn't do what you'd expect: e.g., `null < $x` is `true` for any int but 0; that's because in comparisons between null and something both values are converted to `bool` and compared, with `false < true`).

I wonder if time has finally come to fix this.  I don't know what the right way to do this is, and this diff is meant to sparkle a discussion.  As a starting point, I suggest to allow only comparisons in which either one of the types is dynamic or both types are subtypes of `string`, or `num`, or `DateTime`.

Reviewed By: andrewjkennedy

Differential Revision: D7745607

fbshipit-source-id: a86972cb48595ee7b175f072f0a113055b05024d