Perform Iter to LIter optimization in more cases
Summary:
Perform the Iter to LIter optimization in a few special cases involving
modifications to the array being iterated over. HHBBC can track the local that
is definitely known to be equivalent to the current key in the iteration. If the
local (that is known to be equivalent to the iterator's base) is modified using
that same key, than the Iter to LIter optimization can still be performed.

Iteration can only proceed forward and we've already set the key/value for the
current iteration. We also know that we're modifying the array at the precise
point where we just set the key/value. Therefore this modification cannot
possibly be visible via the iterator and continuing to iterate the modified
array will produce the same results as iterating the original array.

Reviewed By: markw65

Differential Revision: D7845348

fbshipit-source-id: bcf789f0e9619999640fc99f5caf5209b40633d9