sealed classes and interfaces
Summary:
Interfaces and classes can now be shallowly sealed using the `__Sealed` attribute. Sealed classes/interfaces may only be extended/implemented by classes that are explicitly stated as the arguments to the `__Sealed` attribute.

This sealing is single-level such that the classes/interfaces that extend/implement sealed classes/interfaces are not, themselves, implcitly sealed.

Traits are special-cased. Traits cannot be sealed, and when a trait implements a sealed interface, that seal is passed through to the using class, which must be whitelisted.

Reviewed By: oulgen

Differential Revision: D7857001

fbshipit-source-id: c5e0bbc5236868df41d6e460e164395aca1eb761