Pre-smash calls and jumps among Optimize translations, and optimize them on ARM
Summary:
This diff changes the way that optimized translations get published in
retranslate-all mode, allowing the calls and jumps among those translations to
be pre-smashed/linked before the code is published.  Instead of relocating and
publishing each translation at a time, the JIT now follows these steps:

1) Relocate all the optimized translations into the TC.
2) Smash and optimize calls and jumps among optimized translations in place.
3) Publish the metadata for all optimized translations.
4) Publish the code for all optimized translations.

This approach avoids having to later run service requests to smash the jumps and
calls among the optimized translations.  Most importantly, this approach allows
the smashed jumps and calls to be optimized before the code is published.  This
is particularly important for ARM because it allows turning many indirect calls
and jumps into direct ones.  These optimizations are implemented in this diff as
well.

Note that the code is smashed and optimized in place, which leaves some nops
behind.  We may want to run another relocation step across all the optimized
translations between steps 2) and 3) to eliminate these nops, but this diff does
not implement that.

Reviewed By: markw65

Differential Revision: D7859834

fbshipit-source-id: e1aa95bc97005555d959d0beb6e7cc521e6f0bc6