Fix free-on-wrong-thread in async mysql.
Summary:
AsyncMysqlClient::connectAndQuery was capturing an Array in a C++ closure. That closure gets destroyed on a diffrent thread, so the Array gets destroyed on a different thread. If the Array was holding the only reference to its ArrayData, the ArrayData gets freed on a different thread from the one it was allocated on and this causes HHVM's heap to be corrupted.
Fix it by doing the transformation from the HHVM request heap allocated Array to a C++ allocated std::vector before constructing the closure, and closing over the value that is safe to free on any thread instead.
Also fix a leak in AsyncMysqlConnectAndMultiQueryEvent::unserialize where we were taking owership of an ArrayData with Array::detach and then immediately increfing it with cellDup.

Reviewed By: markw65

Differential Revision: D7898564

fbshipit-source-id: 3e28d53b00bd4dfa5d6f200aa9ee9adfec98d00e