Add HH\Rx\mutable function
Summary:
As in title:
Previously results of function calls annotated with `__MutableReturn` or `new` expressions were considered mutably owned.
```
$a = new B(); // $a is mutable
```
With this diff default behavior for these cases will be to treat `$a` as immutable (it is safe since we know that right hand side yields a unique reference to value. In order to make `$a` mutable one should use `HH\Rx\mutable` function - it will indicate that mutability of the value should be preserved (safe for the reason outlined above)

Reviewed By: KendallHopkins

Differential Revision: D7854582

fbshipit-source-id: a0b3b8dfb15b78586c487286200ceb1974caeb9c