Localize wildcards to AKgenerics
Summary:
Previously, we were localizing wildcards as class types. This presented a problem when using them in hints for is-expressions as follows:
```
abstract class C<T as string> {}

function foo(mixed $x): bool {
  return $x is C<_>;
}
```
Here, we would try to subtype the hint to make sure the constraints were satisfied, but... they weren't! A string is incompatible with some random class "_".

In this diff, we localize the wildcard to a fresh type parameter with the constraints automatically applied. This is effectively the same behavior as safe `instanceof`. This eliminates the constraint violation errors in the tests!

sogood

Reviewed By: jamesjwu

Differential Revision: D7887298

fbshipit-source-id: 72530096a0bdc71eb2d4248ce066a53531e0ca3c