Fix gdbprint
Summary:
Casting to a type such as (HPHP::Class) only works if you're in a c++
frame; to guarantee it works regardless, you need to quote it. But
quoting a type like HPHP::Class* will break it if you include the *,
so we use a regex to try to find the right bit to quote. The regex
didn't work for template types though, which broke `lookup func`,
since the type name is something like

  std::__atomic_base<HPHP::Func const*>::__pointer_type

This updates the regex to handle template arguments.

Reviewed By: mxw

Differential Revision: D7946142

fbshipit-source-id: 146d60252febb35eab1bf41889c3828d1e2ee2b8