Adding treadmill admin port command
Summary: When there is a large pending deletes via treadmill size due to a halted treadmill (another symptom is a high hhvm.treadmill.age counter) we have no way to check what is holding the treadmill. The common cases are a debugger that is open and a php request that is not ending but there are other "clients" that can halt the treadmill. This diff instruments all the callsites to startRequest and exposes it via a new admin port command called "treadmill" so we can differentiate which one is responsible.

Reviewed By: markw65

Differential Revision: D7884852

fbshipit-source-id: 947bac254017eb6c47dc4902bf8c7f03e205e385