Match is_resource behavior in IsTypeStruct
Summary:
To ease the migration to the new feature, `$x is resource` should match `is_resource($x)`. This means it's not a simple type-tag check; it also checks if the resource is valid.

This means we have to remove optimizations, but these optimizations never existed for `is_resource`, since it's not terribly hot.

Reviewed By: oulgen

Differential Revision: D7944793

fbshipit-source-id: 5a979ab9a0ace823d24b3d9c5bf63fe6826abb35