Use PGO to decide between bindCall and funcPrologueRedispatch
Summary:
For calls to unknown targets, instead of always binding them to the first callee
seen, this diff uses PGO to decide whether to bind the call to the first callee
or to call the funcPrologueRedispatch unique stub in optimized translations.
This decision is based on how biased towards a single callee the call turns out
to be during profiling. If the call is biased (above
Eval.JitPGOBindCallThreshold, currently set to 50%), we bind a call as before.
Otherwise, we were likely to fail the func guard and then jump to the
funcPrologueRedispatch unique stub in the previous scheme, so now we directly
call that stub.

Reviewed By: markw65

Differential Revision: D7946961

fbshipit-source-id: 9082530230c3899b7291856df8154e0629875539