Fix spec tests in --compare-hh-codegen mode
Summary: D7897001 added an option to the hhvm command in --compare-hh-codegen mode: -vEval.EnableIsExprPrimitiveMigration. This was malformed, and when we were runnign .skipif files, we go the output "Possible bad config node: Eval.EnableIsExprPrimitiveMigration". As a result (because there is a toplevel skipif file) 100% of the spec tests were being skipped in --compare-hh-codegen mode. Fix the option name, now only 20 spec tests skip in --compare-hh-codegen mode (same as normal run mode).

Reviewed By: oulgen, markw65

Differential Revision: D7966388

fbshipit-source-id: 8590a7f7ce3f8ecb6e7f76af5adf17d1c76bb1dd