Add an option to log guard type frequencies to StructLog
Summary:
I wanted this data to guide some changes to DataType that I'm working
on, but it could be more generally useful. I started thinking about just guards
but ended up capturing all calls to emitTypeTest() and all checks against
KindOfRefcountThreshold, since those are the events that will be affected by
any changes to the layout of DataType.

Reviewed By: ricklavoie

Differential Revision: D7841989

fbshipit-source-id: 0b2e960dc041f452c8bf14fe2b4047d5bbf56881