Don't ignore .strict flag when comparing hhas without semdiff
Summary:
We have fixed the .strict flag issues, so stop ignoring it in --no-semdiff.
I also had to fix the hhvm_cmd that we use to dump_hhas_to_temp on the second pass (it needs the original test name as well as the .hhas that we're round-tripping so that it can find relevant .ini files), which in turn revealed a bug in dump_hhas_cmd in that we weren't forcing -vEval.AllowHhas=true even though we use that command with hhas inputs.

Reviewed By: oulgen

Differential Revision: D7983718

fbshipit-source-id: ed906645e73f644fea020344ca447c9d0a3b775a