Remove block of block in parsing
Summary:
In Ast, there are some nodes which include a `block` (`stmt list`); however, due to the way parsing is done, a `Block` (of type `stmt`) is parsed to it becomes the single element of the list. This is inefficient and unnecessary. This commit makes the `block` the actual content of the `Block` instead of a single `Block` `stmt`.

Alternative way: Replace `block` with `stmt` in Ast nodes. I don't see this alternative should be preferred.

Reviewed By: holzensp

Differential Revision: D7729004

fbshipit-source-id: a22ce3ee6163903c6b24ec2fe9b41696d335f5bd