Remove ServerInferType.returned_type_at_pos
Summary:
The function `ServerInferType.returned_type_at_pos` was added to work around the inconvenient formatting of function types in Nuclide. When hovering over the token `getInt`:

```
$obj->getInt();
        ^
```

We would use `returned_type_at_pos` to produce `int` rather than the type of the smallest expression under the cursor (which is `$obj->getInt`, of type `(function(): int)`). Since these function types can be very long, and at the time were displayed in a single-line datatip which was difficult to scroll horizontally, it was much more useful to show the return type than the function type, so we did (D6436228). This also matched with the behavior of hover before switching from typing hooks to the TAST.

Unfortunately, there is a bug in this change. When the smallest expression being hovered over is a function value which is //not// being applied, and occurs inside an expression which //is// a function being applied (as often happens in method chaining), we would incorrectly show the return type of the function value.

Here's an example:

```
($obj->map((int $x) ==> $x * $x))->getValue()
```

When querying for the type of the lambda, we would expect to get `(function(int): int)`. However, since the position we are querying is within a function expression being applied (that expression being `(...)->getValue`), the `returned_type_at_pos` behavior kicks in, incorrectly expecting the resulting function type to represent the entirety of the parenthesized expression (rather than the smallest expression within it under the cursor), and therefore mapping the function type to its return type, resulting in `int`.

We could fix this behavior, but the need for `returned_type_at_pos` is dubious now that IDE hover no longer uses it. Instead, this change removes it altogether, causing --type-at-pos and --type-at-pos-batch to simply return the type of the smallest expression under the cursor (just as IDE hover does).

Reviewed By: pittsw

Differential Revision: D7942984

fbshipit-source-id: d4923c320534057072ca1f93bc115fe21ceace8a