Refactor interruptions
Summary:
Turns out that “here is a list of FDs, enjoy” approach to interrupting is not the most convenient one:

* Watchman FD can change during arbitrary call (there are cases when we handle reconnecting transparently)
* when clients connect, we need to add them to list of monitored FDs in the middle of the job
* similarly, there are many places where client can disconnect and we need to stop monitoring it
* we need to manually track what FD means which and what function needs to be called

Because of this, I change the handler from being statically configured before the job, to being dynamically built based on the environment using provided builder function. The function is called again after operations that can change the environment. The API also changes to just automatically dispatch to the right function when FD is ready instead of passing in the list of all ready fds.

Reviewed By: alexchow

Differential Revision: D7923522

fbshipit-source-id: f819692ef4cab20701ffdebdac4fd6aaa45fc1a5