Cleanup in ServerTypeCheck
Summary:
Long time ago this function would always update majority of fields of this record, so always constructing it from the scratch was fine. Now it's just an unnecessary repetition and boilerplate that is unlikely to save anyone from forgetting to update a field, while making the whole thing less readable.

(Note: this ignores all push blocking failures!)

Reviewed By: alexchow

Differential Revision: D7923455

fbshipit-source-id: 2256b72e90c4a47be07f9b8e4f6059efc0ede63e