Update env.init_env.needs_full_init correctly in presence of interruptions
Summary:
`env.init_env.needs_full_init` should have been updated as part of D7550953, similarly to how `env.full_check` was.

This affects only logging and `cgmem` script (because it reads the logs looking for Server is READY message), but it also gated only to Hack team, so not much harm done because it was broken.

Reviewed By: alexchow

Differential Revision: D7923428

fbshipit-source-id: 10be0529cb1d0747b7a1bdb78298a014aac31f89