Integration test showing that LSP connection is blocked on typechecking
Summary:
This is a beginning of a stack of diffs that will make it possible to handle persistent client connections and requests during typechecking.

For now, it's impossible to even connect to the server during this time. I illustrate it in this test by re-using `hh_loop_forever` machinery added in D7550962 to get the server busy and testing that there is no response to initialize request. This test will evolve to show what things can be done during typechecking (first update in D7923643).

Reviewed By: alexchow

Differential Revision: D7923548

fbshipit-source-id: 487bb77d4d899eab725e45733f0b3bc847ac7d7c