Updates to profiling
Summary:
So the runner I introduced in the last diff is good for getting a general idea for memory and time usage, but it's not as good for profiling: we start up and shut down hh_single_parse/hh_parse every single time, which means the majority of our profiling data is just this initialization.

This diff adds hh_single_parse --benchmark_batch, which allows you to benchmark running a list of files with the ffp. This initializes shared memory once, and then runs every file x times, where x is an argument passed in.

Reviewed By: arxanas

Differential Revision: D7999232

fbshipit-source-id: a89de9229a722efbd62ad9dcd6bf7896c89a0663