not checking tombstone when releasing mixed arrays
Summary:
When creating the tombstone, we make it appear to contain uncounted data.
This saves a branch in release.

This assumes that it is safe to do tvDecRefGen() on a TypedValue with KInvalidDataType.
In current implementation, kInvalidDataType will be considered persistent, so nothing happens.

Also, it appears that we didn't decref the string key being erased.  Fix it as well.

Reviewed By: markw65

Differential Revision: D8015044

fbshipit-source-id: 87ddab405049d4ae595a0df1a5148cdb8f546278