some cleanup in mallctl helpers
Summary:
Make errOK a template parameter as it is always known at compile time.
If USE_JEMALLOC is defined, make it a linking error if the binary isn't linked against libjemalloc.
(what's the point of defining USE_JEMALLOC but do not link with it anyway?)
Also, update some ifdefs now that jemalloc 5.1 is out.

Reviewed By: interwq

Differential Revision: D8020833

fbshipit-source-id: bcec5556c357bf0666eb0b9740d175602b8e6d7d