Ability to suspend handling of clients if they need full rechek or want to modify server state
Summary:
To avoid diverging code too much, I want to re-use `ServerMain.handle_connection` during interruptions, because it will be mostly identical. The main difference will be handling of methods that need full recheck or want to change server state - we'll sometimes immediately handle them, and sometimes want to enqueue them for later.

I achieve it by returning a variant with an updated environment, or a continuation that can be finished later. In existing callsites, always running the continuation to completion.

Reviewed By: alexchow

Differential Revision: D7923599

fbshipit-source-id: 459601d225bda640402dace11ec1dbf9e15c0746