SUBSCRIBE_DIAGNOSTICS doesn't require full check
Summary: Even though we can now connect in non blocking way, the first things that basically any client will do is subscribe diagnostics which is unfortunately blocking. It doesn't need to be, and was done so mostly for simplicity, before we had incrementally updated error map, and because there was no benefit to it (since all the other things are blocking anyway).

Reviewed By: alexchow

Differential Revision: D7923668

fbshipit-source-id: 4e823db6557ecb014f0885d96d8f57731c41522e