Add an option to control the number of workers in Serdes mode.
Summary:
In deserialize mode, we normally override the preset number of worker
threads, because we want to fully utilize the cpu. This gives us an
option to override it (this is mainly useful for setting it to 1 for
debugability). If we were in DeserializeOrGenerate mode, we also
override the thread count for serialize (again, mostly for debugging,
to unjumble any TRACE output, and match the deserialize behavior).

Reviewed By: ottoni

Differential Revision: D8029669

fbshipit-source-id: d4d4eaac6dc935335be26fdc910eaa889850b86c