Improve typing of shapes with optional fields
Summary:
Subtyping of shapes with optional fields is broken.

Consider the following example code:

```
<?hh // strict

function type_hole<T>(shape(...) $s): ?T {
  return Shapes::idx($s, 'x');
}

class A {}
class B {
  public function f(): void {}
}

function test(): void {
  $x = type_hole(shape('x' => new A()));
  if ($x !== null) {
    $x->f();
  }
}
```
It type-checks but unsurprisingly a call to `test()` fatals at runtime.

What does it have to do with subtyping of shapes with optional fields?  Type-checking the call `Shapes::idx($s, 'x')` involves constructing a "fake" shape type `shape(?'x' => ?#n[unresolved])` and checking that the type of `$s` is a subtype of that "fake" shape type.  It is this subtyping check that is broken: it succeeds without constraining the type variable `#n`, so that the inferred type of the expression `Shapes::idx($s, 'x')` becomes `?#n[unresolved]`, which is compatible with `?T`.  (Actually, the current implementation has some special logic for the case when one tries to access a field of an open shape that hasn't been explicitly unset:

diffusion/FBS/browse/master/fbcode/hphp/hack/src/typing/typing_shapes.ml;b55d900dd87a3f9b3ba4743f31534fb523e798cc$61-82

but that logic doesn't apply when the type of the argument to `Shapes::idx()` is an unresolved shape, which is easily achieved in the above example.)

Here is another example showing that something is off:

```
<?hh // strict

type s = shape(...);
type t = shape(?'z' => mixed, ...);

function test(s $s): t {
  return $s;
}
```

This fails to type-check: the type-checker is complaining that the field `'z'` may be set to an unknown type in `$s`, which is a shape of type `s`.  I think this is not correct: if `$s` doesn't have a field `'z'`, then its type is compatible with `t` (because in `t`, `z` is declared optional); and if `$s` has a field `'z'`, then the type of the field in `$s` is a subtype of the field `'z'` in `t`, `mixed`.  This should type-check.

To summarize, the problem is in how we deal with missing optional fields during subtyping and unification.  The current implementation errors out if we have an optional field declared in the supertype that is not declared and not explicitly removed (using `Shapes::removeKey()`) in the subtype.  The correct implementation, I think, is to require that in this case the type of the field declared in the supertype is (a supertype of) `mixed`.

There is a flag `--disable_optional_and_unknown_shape_fields`, which is not applied to www but which exists in the open-source version of Hack for migration purposes.  When invoked, it results, in particular, in `shape()` being treated as `shape(...)` and `shape('z' => ?t)` as `shape(?'z' => ?t)` (i.e., all shapes are open and nullable fields are considered optional; see also [[ https://hhvm.com/blog/2017/09/26/hhvm-3-22.html#nullable-vs-optional-fields-in-shapes | this ]] page).  The above reasoning should apply in this case as well.  In particular, `shape()` should be a subtype of `shape('z' => mixed)`.

Reviewed By: andrewjkennedy

Differential Revision: D7929729

fbshipit-source-id: c4e0d8b85b2e0cf9adc617bbf61bb46663721fa3