Shapes::keyExists() should refine its first argument's type when used as a condition
Summary:
A call to `Shapes::keyExists()` used as a condition should refine the type of its first argument.

More precisely, `Shapes::keyExists($s, 'x')` should generate a lint warning that this expression is always true if the type of `$s` is a shape that has a required field `'x'` and is always false if it is a closed shape that doesn't contain `'x'` or a shape about which it is known that `'x'` has been unset.  Such a warning is easier to generate by analyzing the typed AST, so it's not implemented in this diff.  I also think it shouldn't be a type error to call `Shapes::keyExists($s, 'x')` if `$s` is known not to have a field `'x'`.

Otherwise, if `Shapes::keyExists($s, 'x')` is used as a condition, it should refine the type of `$s` as follows:
```
if (Shapes::keyExists($s, 'x')) {
  // the type of $s should be:
  // - shape('x' => mixed, <fields>, ...)
  //   if the incoming type of $s is shape(<fields>, ...) where <fields> doesn't contain 'x'
  // - shape('x' => t, <fields>)
  //   if the incoming type of $s is shape(?'x' => t, <fields>), where <fields> may or may not include ...
}
```

Reviewed By: andrewjkennedy

Differential Revision: D8009676

fbshipit-source-id: 5d64a78a44906df7ff27a3537b15b51d2d30b871